initial

+++ client/src/components/Body/ProfilePage.jsx
@@ -74,17 +74,23 @@ const ProfilePage = ({ telegramId }) => {
         const confirmDelete = window.confirm("Вы уверены, что хотите удалить свой профиль?");  
         if (confirmDelete) {  
             try {  
-                const response = await fetch(`https://strology.vercel.app/api/users/${user.telegramId}`, {  // Используйте telegramId  
+                const userCheckResponse = await axios.get(`https://strology.vercel.app/api/users/${user.telegramId}`);  
+                if (!userCheckResponse.data.exists) {  
+                    alert("Пользователь не найден. Невозможно удалить профиль.");  
+                    return;  
+                }  
+    
+                // Удаление пользователя  
+                const response = await fetch(`https://strology.vercel.app/api/users/${user.telegramId}`, {  
                     method: 'DELETE',  
                     headers: {  
                         'Content-Type': 'application/json',  
-                        // Добавьте здесь токен авторизации, если требуется  
                     },  
                 });  
     
                 if (response.ok) {  
                     alert("Профиль успешно удален!");  
-                    navigate('/'); // Перенаправьте пользователя на главную или страницу входа  
+                    navigate('/'); // Перенаправление  
                 } else {  
                     const errorData = await response.json();  
                     alert("Ошибка при удалении профиля: " + errorData.message);
